Add <ScanDir> as an optional keyword when defining directory in lima_conf (#17)
The main purpose of this pull request is to allow saving the images to directories outside {ScanDir} (e.g. to allow saving to /tmp/... or any other absolute paths). This is useful when the computer that saves the file(s) has specific mount points not accessible from the Sardana machine.